requirements.txt
# GlamAI Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0

# Azure AI Services
azure-ai-vision-imageanalysis==1.0.0b1
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-speech==1.34.0
azure-ai-textanalytics==5.3.0
azure-ai-translation-text==1.0.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-ai-formrecognizer==3.3.2

# OpenAI for LLM
openai==1.10.0

# ML & Computer Vision
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
scikit-learn==1.4.0
# face-recognition==1.3.0
mediapipe==0.10.9

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
openpyxl==3.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.2.0
pydantic-extra-types==2.5.0

# Background Tasks & Cachingf
celery==5.3.6
redis==5.0.1
aioredis==2.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0


# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Environment
python-dotenv==1.0.0

# Image Processing
imgaug==0.4.0

# Server (Production)
gunicorn==21.2.0